* fixed handling of rapid bank select and program change messages sent   to the same sampler channel (patch from the Open Octave project,   slightly adjusted)
git-svn-id: https://svn.linuxsampler.org/svn@2277 d2c4447f-3961-4fff-bec3-96d84be9abe5

+++ linuxsampler/trunk/src/engines/EngineChannel.h
@@ -1,6 +1,6 @@
 /***************************************************************************
  *                                                                         *
- *   Copyright (C) 2005 - 2010 Christian Schoenebeck                       *
+ *   Copyright (C) 2005 - 2011 Christian Schoenebeck                       *
  *                                                                         *
  *   This program is free software; you can redistribute it and/or modify  *
  *   it under the terms of the GNU General Public License as published by  *
@@ -346,7 +346,7 @@ namespace LinuxSampler {
              *
              * This method is not real-time safe.
              */
-            void ExecuteProgramChange(uint8_t Program);
+            void ExecuteProgramChange(uint32_t Program);
 
         protected:
             EngineChannel();
